# src/ConsumerDataStandards/Controllers/PayeesApi.cs
/*
 * Consumer Data Standards
 *
 * API sets created by the Australian Consumer Data Standards to meet the needs of the Consumer Data Right
 *
 * OpenAPI spec version: 1
 * 
 * Generated by: https://github.com/swagger-api/swagger-codegen.git
 */

using System;
using System.Collections.Generic;
using System.Linq;
using System.Net;
using System.Threading.Tasks;
using Microsoft.AspNetCore.Http;
using Microsoft.AspNetCore.Mvc;
using Microsoft.AspNetCore.WebUtilities;
using Microsoft.Extensions.Logging;
using Microsoft.Extensions.Primitives;
using Swashbuckle.AspNetCore.SwaggerGen;
using Newtonsoft.Json;
using System.ComponentModel.DataAnnotations;
using ConsumerDataStandards.Attributes;
using ConsumerDataStandards.Models;

namespace ConsumerDataStandards.Controllers
{ 
    /// <summary>
    /// 
    /// </summary>
    public class PayeesApiController : Controller
    { 
        /// <summary>
        /// Get Payee Detail
        /// </summary>
        /// <remarks>Obtain detailed information on a single payee</remarks>
        /// <param name="payeeId">The ID used to locate the details of a particular payee</param>
        /// <response code="200">Success</response>
        [HttpGet]
        [Route("/cds-au/v1/banking/payees/{payeeId}")]
        [ValidateModelState]
        [SwaggerOperation("GetPayeeDetail")]
        [SwaggerResponse(statusCode: 200, type: typeof(ResponseBankingPayeeById), description: "Success")]
        public virtual IActionResult GetPayeeDetail([FromRoute][Required]string payeeId)
        { 
            //TODO: Uncomment the next line to return response 200 or use other options such as return this.NotFound(), return this.BadRequest(..), ...
            // return StatusCode(200, default(ResponseBankingPayeeById));

            string exampleJson = null;
            exampleJson = "{\n  \"data\" : \"\",\n  \"meta\" : { },\n  \"links\" : {\n    \"self\" : \"self\"\n  }\n}";
            
            var example = exampleJson != null
            ? JsonConvert.DeserializeObject<ResponseBankingPayeeById>(exampleJson)
            : default(ResponseBankingPayeeById);
            //TODO: Change the data returned
            return new ObjectResult(example);
        }

        /// <summary>
        /// Get Payees
        /// </summary>
        /// <remarks>Obtain a list of pre-registered payees</remarks>
        /// <param name="type">Filter on the payee type field.  In addition to normal type field values, ALL can be specified to retrieve all payees.  If absent the assumed value is ALL</param>
        /// <param name="page">Page of results to request (standard pagination)</param>
        /// <param name="pageSize">Page size to request. Default is 25 (standard pagination)</param>
        /// <response code="200">Success</response>
        [HttpGet]
        [Route("/cds-au/v1/banking/payees")]
        [ValidateModelState]
        [SwaggerOperation("ListPayees")]
        [SwaggerResponse(statusCode: 200, type: typeof(ResponseBankingPayeeList), description: "Success")]
        public virtual IActionResult ListPayees([FromQuery]string type, [FromQuery]int? page, [FromQuery]int? pageSize)
        { 
            //TODO: Uncomment the next line to return response 200 or use other options such as return this.NotFound(), return this.BadRequest(..), ...
            // return StatusCode(200, default(ResponseBankingPayeeList));

            string exampleJson = null;
            exampleJson = "{\n  \"data\" : {\n    \"payees\" : [ {\n      \"nickname\" : \"nickname\",\n      \"description\" : \"description\",\n      \"payeeId\" : \"payeeId\",\n      \"type\" : \"DOMESTIC\",\n      \"creationDate\" : \"creationDate\"\n    }, {\n      \"nickname\" : \"nickname\",\n      \"description\" : \"description\",\n      \"payeeId\" : \"payeeId\",\n      \"type\" : \"DOMESTIC\",\n      \"creationDate\" : \"creationDate\"\n    } ]\n  },\n  \"meta\" : {\n    \"totalRecords\" : 0,\n    \"totalPages\" : 6\n  },\n  \"links\" : {\n    \"next\" : \"next\",\n    \"last\" : \"last\",\n    \"prev\" : \"prev\",\n    \"self\" : \"self\",\n    \"first\" : \"first\"\n  }\n}";
            
            var example = exampleJson != null
            ? JsonConvert.DeserializeObject<ResponseBankingPayeeList>(exampleJson)
            : default(ResponseBankingPayeeList);
            //TODO: Change the data returned
            return new ObjectResult(example);
        }
    }
}